Code changes to enable Dual writes.
Summary:
The goal of this change is to enable SQL First dual writes as first priority.
The change is now incomplete and just for review.

Test Plan: fullcycle

Reviewers: haomin, txchen

Reviewed By: txchen

Subscribers: kg

Differential Revision: https://code.silkcloud.info/D39

+++ cart/cart-db/src/main/groovy/com/junbo/cart/db/entity/CartEntity.java
@@ -28,7 +28,7 @@ public class CartEntity {
 
     private String cartName;
 
-    private Long resourceAge;
+    private Integer resourceAge;
 
     private String properties;
 
@@ -84,11 +84,11 @@ public class CartEntity {
 
     @Column(name = "RESOURCE_AGE")
     @Version
-    public Long getResourceAge() {
+    public Integer getResourceAge() {
         return resourceAge;
     }
 
-    public void setResourceAge(Long resourceAge) {
+    public void setResourceAge(Integer resourceAge) {
         this.resourceAge = resourceAge;
     }
